Data types can be hard
I was using an int var to store the delay time which works fine for
small delay times but being that delay is in miliseconds  can become
very large when using minutes.

Switched from int to unsigned long

+++ NoDelay.cpp
@@ -33,16 +33,17 @@
 /**************************************************************************/
 noDelay::noDelay(void)
 {
+	
 	preMills = 0;
 	setdelay(0);
 }
-noDelay::noDelay(int dtime)
+noDelay::noDelay(unsigned long dtime)
 {
 	preMills = 0;
 	setdelay(dtime);
 }
 
-noDelay::noDelay(int dtime, funTocall funcall)
+noDelay::noDelay(unsigned long dtime, funTocall funcall)
 {
 	preMills = 0;
 	setdelay(dtime);
